backend/requirements.txt
django>=4.2.0,<5.0
strawberry-graphql-django>=0.47.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
